# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/catalog/games/{id}/fields": {
      "post": {
        "operationId": "CatalogItemsAdminController_createField",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInputField"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Last in the order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputField"
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/admin/catalog/games/{id}/fields/order": {
      "put": {
        "operationId": "CatalogItemsAdminController_reorderFields",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reorder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "In their new order",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InputField"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/admin/catalog/fields/{id}": {
      "patch": {
        "operationId": "CatalogItemsAdminController_updateField",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInputField"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The field",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputField"
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/admin/catalog/fields/{id}/archive": {
      "post": {
        "operationId": "CatalogItemsAdminController_archiveField",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The field",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputField"
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/admin/catalog/fields/{id}/restore": {
      "post": {
        "operationId": "CatalogItemsAdminController_restoreField",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The field",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputField"
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/admin/catalog/games/{id}/products": {
      "post": {
        "operationId": "CatalogItemsAdminController_createProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProduct"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Last in the order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/admin/catalog/games/{id}/products/order": {
      "put": {
        "operationId": "CatalogItemsAdminController_reorderProducts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Reorder"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "In their new order",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/admin/catalog/products/{id}": {
      "patch": {
        "operationId": "CatalogItemsAdminController_updateProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProduct"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/admin/catalog/products/{id}/archive": {
      "post": {
        "operationId": "CatalogItemsAdminController_archiveProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/admin/catalog/products/{id}/restore": {
      "post": {
        "operationId": "CatalogItemsAdminController_restoreProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vertex Digital API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Reorder": {
        "type": "object",
        "properties": {
          "ids": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
...
      "CatalogStatus": {
        "type": "string",
        "enum": [
          "active",
          "paused"
        ]
      },
...
      "InputField": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "sortOrder": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "archivedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "gameId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "key": {
            "type": "string"
          },
          "labelAr": {
            "type": "string"
          },
          "helpAr": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/InputFieldType"
          },
          "required": {
            "type": "boolean"
          },
          "minLength": {
            "nullable": true,
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "maxLength": {
            "nullable": true,
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "options": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectOption"
            }
          }
        },
        "required": [
          "id",
          "sortOrder",
          "archivedAt",
          "createdAt",
          "updatedAt",
          "gameId",
          "key",
          "labelAr",
          "helpAr",
          "type",
          "required",
          "minLength",
          "maxLength",
          "options"
        ],
        "additionalProperties": false
      },
      "InputFieldType": {
        "type": "string",
        "enum": [
          "digits",
          "text",
          "select",
          "phone"
        ]
      },
      "SelectOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "pattern": "^[a-z0-9_-]{1,32}$"
          },
          "labelAr": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "pattern": "^[^\\p{Cc}]+$"
          }
        },
        "required": [
          "value",
          "labelAr"
        ],
        "additionalProperties": false
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "sortOrder": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "archivedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "gameId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "kind": {
            "$ref": "#/components/schemas/ProductKind"
          },
          "nameAr": {
            "type": "string"
          },
          "gameAmount": {
            "nullable": true,
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "officialPriceUsdUnits": {
            "nullable": true,
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "maxQuantity": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "regionAr": {
            "nullable": true,
            "type": "string"
          },
          "redemptionAr": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CatalogStatus"
          },
          "availability": {
            "$ref": "#/components/schemas/ProductAvailability"
          }
        },
        "required": [
          "id",
          "sortOrder",
          "archivedAt",
          "createdAt",
          "updatedAt",
          "gameId",
          "kind",
          "nameAr",
          "gameAmount",
          "officialPriceUsdUnits",
          "maxQuantity",
          "regionAr",
          "redemptionAr",
          "status",
          "availability"
        ],
        "additionalProperties": false
      },
      "ProductKind": {
        "type": "string",
        "enum": [
          "direct",
          "code"
        ]
      },
      "ProductAvailability": {
        "type": "string",
        "enum": [
          "hidden",
          "paused",
          "paused_by_margin_guard",
          "out_of_stock",
          "available"
        ]
      },
...
      "CreateInputField": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "key": {
                "type": "string",
                "pattern": "^[a-z][a-z0-9_]{1,31}$"
              },
              "labelAr": {
                "type": "string",
                "minLength": 1,
                "maxLength": 40,
                "pattern": "^[^\\p{Cc}]+$"
              },
              "helpAr": {
                "nullable": true,
                "type": "string",
                "minLength": 1,
                "maxLength": 200,
                "pattern": "^[^\\p{Cc}]+$"
              },
              "required": {
                "type": "boolean"
              },
              "type": {
                "type": "string",
                "enum": [
                  "digits"
                ]
              },
              "minLength": {
                "default": null,
                "nullable": true,
                "type": "integer",
                "minimum": 1,
                "maximum": 32
              },
              "maxLength": {
                "default": null,
                "nullable": true,
                "type": "integer",
                "minimum": 1,
                "maximum": 32
              },
              "options": {
                "default": null,
                "type": "string",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            },
            "required": [
              "key",
              "labelAr",
              "required",
              "type"
            ]
          },
          {
            "type": "object",
            "properties": {
              "key": {
                "type": "string",
                "pattern": "^[a-z][a-z0-9_]{1,31}$"
              },
              "labelAr": {
                "type": "string",
                "minLength": 1,
                "maxLength": 40,
                "pattern": "^[^\\p{Cc}]+$"
              },
              "helpAr": {
                "nullable": true,
                "type": "string",
                "minLength": 1,
                "maxLength": 200,
                "pattern": "^[^\\p{Cc}]+$"
              },
              "required": {
                "type": "boolean"
              },
              "type": {
                "type": "string",
                "enum": [
                  "text"
                ]
              },
              "minLength": {
                "default": null,
                "nullable": true,
                "type": "integer",
                "minimum": 1,
                "maximum": 64
              },
              "maxLength": {
                "default": null,
                "nullable": true,
                "type": "integer",
                "minimum": 1,
                "maximum": 64
              },
              "options": {
                "default": null,
                "type": "string",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            },
            "required": [
              "key",
              "labelAr",
              "required",
              "type"
            ]
          },
          {
            "type": "object",
            "properties": {
              "key": {
                "type": "string",
                "pattern": "^[a-z][a-z0-9_]{1,31}$"
              },
              "labelAr": {
                "type": "string",
                "minLength": 1,
                "maxLength": 40,
                "pattern": "^[^\\p{Cc}]+$"
              },
              "helpAr": {
                "nullable": true,
                "type": "string",
                "minLength": 1,
                "maxLength": 200,
                "pattern": "^[^\\p{Cc}]+$"
              },
              "required": {
                "type": "boolean"
              },
              "type": {
                "type": "string",
                "enum": [
                  "select"
                ]
              },
              "minLength": {
                "default": null,
                "type": "string",
                "nullable": true,
                "enum": [
                  null
                ]
              },
              "maxLength": {
                "default": null,
                "type": "string",
                "nullable": true,
                "enum": [
                  null
                ]
              },
              "options": {
                "minItems": 2,
                "maxItems": 50,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SelectOption"
                }
              }
            },
            "required": [
              "key",
              "labelAr",
              "required",
              "type",
              "options"
            ]
          },
          {
            "type": "object",
            "properties": {
              "key": {
                "type": "string",
                "pattern": "^[a-z][a-z0-9_]{1,31}$"
              },
              "labelAr": {
                "type": "string",
                "minLength": 1,
                "maxLength": 40,
                "pattern": "^[^\\p{Cc}]+$"
              },
              "helpAr": {
                "nullable": true,
                "type": "string",
                "minLength": 1,
                "maxLength": 200,
                "pattern": "^[^\\p{Cc}]+$"
              },
              "required": {
                "type": "boolean"
              },
              "type": {
                "type": "string",
                "enum": [
                  "phone"
                ]
              },
              "minLength": {
                "default": null,
                "type": "string",
                "nullable": true,
                "enum": [
                  null
                ]
              },
              "maxLength": {
                "default": null,
                "type": "string",
                "nullable": true,
                "enum": [
                  null
                ]
              },
              "options": {
                "default": null,
                "type": "string",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            },
            "required": [
              "key",
              "labelAr",
              "required",
              "type"
            ]
          }
        ]
      },
      "UpdateInputField": {
        "type": "object",
        "properties": {
          "labelAr": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "pattern": "^[^\\p{Cc}]+$"
          },
          "helpAr": {
            "nullable": true,
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "pattern": "^[^\\p{Cc}]+$"
          },
          "required": {
            "type": "boolean"
          },
          "minLength": {
            "nullable": true,
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "maxLength": {
            "nullable": true,
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "options": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectOption"
            }
          }
        }
      },
      "CreateProduct": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "direct"
                ]
              },
              "nameAr": {
                "type": "string",
                "minLength": 1,
                "maxLength": 60,
                "pattern": "^[^\\p{Cc}]+$"
              },
              "gameAmount": {
                "nullable": true,
                "type": "integer",
                "minimum": 1,
                "maximum": 1000000000
              },
              "officialPriceUsdUnits": {
                "nullable": true,
                "type": "integer",
                "minimum": 1,
                "maximum": 10000000000
              },
              "maxQuantity": {
                "default": 1,
                "type": "integer",
                "minimum": 1,
                "maximum": 50
              }
            },
            "required": [
              "kind",
              "nameAr"
            ]
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "code"
                ]
              },
              "nameAr": {
                "type": "string",
                "minLength": 1,
                "maxLength": 60,
                "pattern": "^[^\\p{Cc}]+$"
              },
              "gameAmount": {
                "nullable": true,
                "type": "integer",
                "minimum": 1,
                "maximum": 1000000000
              },
              "officialPriceUsdUnits": {
                "nullable": true,
                "type": "integer",
                "minimum": 1,
                "maximum": 10000000000
              },
              "maxQuantity": {
                "default": 10,
                "type": "integer",
                "minimum": 1,
                "maximum": 50
              },
              "regionAr": {
                "nullable": true,
                "type": "string",
                "minLength": 1,
                "maxLength": 64,
                "pattern": "^[^\\p{Cc}]+$"
              },
              "redemptionAr": {
                "nullable": true,
                "type": "string",
                "minLength": 1,
                "maxLength": 2000,
                "pattern": "^(?:[^\\p{Cc}]|\\n)+$"
              }
            },
            "required": [
              "kind",
              "nameAr"
            ]
          }
        ]
      },
      "UpdateProduct": {
        "type": "object",
        "properties": {
          "nameAr": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60,
            "pattern": "^[^\\p{Cc}]+$"
          },
          "gameAmount": {
            "nullable": true,
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000000
          },
          "officialPriceUsdUnits": {
            "nullable": true,
            "type": "integer",
            "minimum": 1,
            "maximum": 10000000000
          },
          "maxQuantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50
          },
          "regionAr": {
            "nullable": true,
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "pattern": "^[^\\p{Cc}]+$"
          },
          "redemptionAr": {
            "nullable": true,
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "pattern": "^(?:[^\\p{Cc}]|\\n)+$"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CatalogStatus"
              }
            ]
          }
        }
      },
...
    }
  }
}